Replace GitHub Actions Cloudflare DNS step with an EventBridge+Lambda dns-sync

+++ src/main/java/bflow/subscription/WompiApiClient.java
@@ -90,7 +90,7 @@ public class WompiApiClient {
 
         cachedToken = response.accessToken();
         tokenExpiry = Instant.now().plusSeconds(
-                response.expiresIn() - TOKEN_MARGIN_SECONDS
+                (long) response.expiresIn() - TOKEN_MARGIN_SECONDS
         );
         return cachedToken;
     }
